Move types to existing place: use-site-logs-query
