+++ client/data/hosting/use-site-logs-query.ts
@@ -11,9 +11,32 @@ interface SiteLogsAPIResponse {
 	};
 }
 
+export interface ServerLog {
+	body_bytes_sent: number;
+	cached: string;
+	date: string;
+	http_host: string;
+	http_referer: string;
+	request_type: 'GET' | 'HEAD' | 'POST' | 'PUT' | 'DELETE';
+	request_url: string;
+	status: '200' | '301' | '302' | '400' | '401' | '403' | '404' | '429' | '500';
+	timestamp: number;
+}
+
+export interface PHPLog {
+	atomic_site_id: number;
+	file: string;
+	kind: string;
+	line: number;
+	message: string;
+	name: string;
+	severity: 'User' | 'Warning' | 'Deprecated' | 'Fatal error';
+	timestamp: string;
+}
+
 export type SiteLogsData = {
 	total_results: number;
-	logs: Record< string, unknown >[];
+	logs: Record< string, ServerLog | PHPLog >[];
 	scroll_id: string | null;
 	has_more: boolean;
 };
@@ -83,10 +106,11 @@ export function useSiteLogsQuery(
 		staleTime: Infinity, // The logs within a specified time range never change.
 		select( { data } ) {
 			return {
-				...data,
 				has_more: !! data.scroll_id,
 				total_results:
 					typeof data.total_results === 'number' ? data.total_results : data.total_results.value,
+				logs: data.logs as Record< string, ServerLog | PHPLog >[],
+				scroll_id: data.scroll_id,
 			};
 		},
 		meta: {
