+++ client/sites/tools/logs/components/site-logs-table/index.tsx
@@ -2,8 +2,7 @@ import { sprintf } from '@wordpress/i18n';
 import { useI18n } from '@wordpress/react-i18n';
 import clsx from 'clsx';
 import { memo, useMemo } from 'react';
-import { SiteLogsData } from 'calypso/data/hosting/use-site-logs-query';
-import { LogType } from 'calypso/sites/tools/logs';
+import { SiteLogsData, LogType } from 'calypso/data/hosting/use-site-logs-query';
 import { useCurrentSiteGmtOffset } from 'calypso/sites/tools/logs/hooks/use-current-site-gmt-offset';
 import { useSelector } from 'calypso/state';
 import { getCurrentUserLocale } from 'calypso/state/current-user/selectors';
