+++ client/sites/tools/logs/components/site-logs-toolbar/index.tsx
@@ -9,8 +9,8 @@ import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { useState } from 'react';
 import { useLocalizedMoment } from 'calypso/components/localized-moment';
+import { LogType } from 'calypso/data/hosting/use-site-logs-query';
 import { navigate } from 'calypso/lib/navigate';
-import { LogType } from 'calypso/sites/tools/logs';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { useCurrentSiteGmtOffset } from '../../hooks/use-current-site-gmt-offset';
